Application works to spec. Solution is with making another no fly boundary near the problem area. TODO: Comments, perhaps redundancy when path is not found in any way

+++ aqmaps/src/main/java/uk/ac/ed/inf/aqmaps/SensorData.java
@@ -1,11 +1,10 @@
 package uk.ac.ed.inf.aqmaps;
 
-import java.math.BigDecimal;
 import com.mapbox.geojson.*;
 
 public class SensorData {
 	String location;
-	BigDecimal battery;
+	float battery;
 	String reading;	
 	String lat;
 	String lng;
